moved file TestSupportPlatform.cs
