--- source/DasBlog.Tests/FunctionalTests/Common/Constants.cs
@@ -1,6 +1,0 @@
-namespace DasBlog.Tests.FunctionalTests.Common
-{
-	internal static class Constants
-	{
-	}
-}
